refactor: create match package
